+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.3.0-alpha (Unreleased)
 
+* Added `aftercanvasclear` event.
+
 * `PhSim.Vector.neg` has been added. This creates a vector that is equal in length but opposite in direction of the original vector.
 
 * `PhSim.Vector.extract` has been added. This creates a new vector out of an object `o` that